backend_python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
uuid==1.30
chrono==1.0.0
validators==0.22.0
httpx==0.25.2
aiofiles==23.2.1
